+++ ESFramework/UIKit/UIApplication+ESAdditions.h
@@ -97,7 +97,6 @@ FOUNDATION_EXTERN NSString *const ESAppPreviousVersionUserDefaultsKey;
 @property (nonatomic, readonly) NSString *appVersion;
 @property (nonatomic, readonly) NSString *appBuildVersion;
 @property (nonatomic, readonly) NSString *appFullVersion; // "1.2.4 (210)"
-@property (nonatomic, readonly) BOOL isUIViewControllerBasedStatusBarAppearance;
 
 /**
  * The first filename within the CFBundleIconFiles in the Info.plist file.